# backend/storage/app/scribe/openapi.yaml
openapi: 3.0.3
info:
  title: 'JourneyPlanner API Documentation'
  description: 'JourneyPlanner API - for internal use only'
  version: 1.0.0
servers:
  -
    url: 'http://localhost:8000'
...
components:
  securitySchemes:
    default:
      type: http
      scheme: bearer
      description: 'You can retrieve your token by visiting /token.'
security:
  -
    default: []
...
paths:
...
  /api/journey:
    get:
      summary: 'Get journeys'
      operationId: getJourneys
      description: 'Display all journeys of the authenticated user.'
      parameters: []
      responses:
        200:
          description: Success.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                example:
                  -
                    id: 9c23279f-f2a7-4666-bc00-56dfe0c47342
                    name: '4BI Klassenreise-Stockholm'
                    destination: 'Stockholm, Stockholm, Sweden'
                    mapbox_full_address: 'Stockholm, Stockholm, Sweden'
                    mapbox_id: null
                    from: '2024-05-27T00:00:00.000000Z'
                    to: '2024-06-01T00:00:00.000000Z'
                    role: 1
                    creator: null
                  -
                    id: 9c4bbada-7458-422a-832d-c09794db6013
                    name: Istanbul
                    destination: 'Istanbul, Türkiye'
                    mapbox_full_address: 'Istanbul, Türkiye'
                    mapbox_id: 1234567890abcdefghij
                    from: '2024-08-01T00:00:00.000000Z'
                    to: '2024-08-08T00:00:00.000000Z'
                    role: 1
                    creator: null
                  -
                    id: 9c50ae5b-a3e8-4aee-9ec4-fed26b47a274
                    name: 'Faakersee - JP Teambuilding Reise'
                    destination: 'Egger Seeuferstraße 40, 9580 Villach, Austria'
                    mapbox_full_address: 'Egger Seeuferstraße 40, 9580 Villach, Austria'
                    mapbox_id: 1234567890abcdefghij
                    from: '2024-08-10T00:00:00.000000Z'
                    to: '2024-08-16T00:00:00.000000Z'
                    role: 0
                    creator: null
                  -
                    id: 9c56f79b-f230-456a-be1f-aa7f1a729a17
                    name: 'Interrail - JP Discovers EU'
                    destination: Austria
                    mapbox_full_address: Austria
                    mapbox_id: 1234567890abcdefghij
                    from: '2025-07-01T00:00:00.000000Z'
                    to: '2025-07-31T00:00:00.000000Z'
                    role: 1
                    creator: null
                  -
                    id: 9ce95180-ff68-481d-af61-bf27ebf46f57
                    name: 'Klassenreise Krakau'
                    destination: 'Kraków, Lesser Poland Voivodeship, Poland'
                    mapbox_full_address: 'Kraków, Lesser Poland Voivodeship, Poland'
                    mapbox_id: 1234567890abcdefghij
                    from: '2024-10-01T00:00:00.000000Z'
                    to: '2024-10-04T00:00:00.000000Z'
                    role: 1
                    creator: null
                  -
                    id: 9dbcbefd-13d8-4e2d-bc50-49f34eeaf1fa
                    name: 'Krakau 2.0'
                    destination: 'Kraków, Lesser Poland Voivodeship, Poland'
                    mapbox_full_address: 'Kraków, Lesser Poland Voivodeship, Poland'
                    mapbox_id: 1234567890abcdefghij
                    from: '2024-12-18T00:00:00.000000Z'
                    to: '2024-12-29T00:00:00.000000Z'
                    role: 1
                    creator: null
        403:
          description: Unauthorized.
          content:
            application/json:
              schema:
                type: object
                example:
                  message: 'This action is unauthorized.'
                properties:
                  message:
                    type: string
                    example: 'This action is unauthorized.'
                    description: 'The error message.'
                    enum: []
                required:
                  - message
      tags:
        - Journey
    post:
      summary: 'Store journey'
      operationId: storeJourney
      description: "Store the new journey and add the authenticated user to it.\n\nIf the user isn't authenticated, the journey is created in guest mode."
      parameters: []
      responses:
        201:
          description: Success.
          content:
            application/json:
              schema:
                type: object
                example:
                  message: 'Journey created successfully'
                  journey:
                    name: 'Seine-sational trip'
                    destination: Paris
                    from: '2025-03-23T00:00:00.000000Z'
                    to: '2025-03-30T00:00:00.000000Z'
                    mapbox_full_address: 'Paris, France'
                    mapbox_id: 1234567890abcdefghij
                    invite: bf74e3c0-35d0-4854-9335-6ac6b9ce8956
                    longitude: 2.348392
                    latitude: 48.853495
                    length: 8
                    id: 9e807de1-4b8e-4cbd-9dc5-cf2addb6a89a
                    updated_at: '2025-03-23T16:24:09.000000Z'
                    created_at: '2025-03-23T16:24:09.000000Z'
                    creator: null
                properties:
                  message:
                    type: string
                    example: 'Journey created successfully'
                    description: 'The succes/error message.'
                    enum: []
                  journey:
                    type: object
                    properties:
                      name:
                        type: string
                        example: 'Seine-sational trip'
                        description: 'The name of the created journey.'
                        enum: []
                      destination:
                        type: string
                        example: Paris
                        description: 'The destination of the created journey.'
                        enum: []
                      from:
                        type: string
                        example: '2025-03-23T00:00:00.000000Z'
                        description: 'The start date of the created journey.'
                        enum: []
                      to:
                        type: string
                        example: '2025-03-30T00:00:00.000000Z'
                        description: 'The end date of the created journey.'
                        enum: []
                      mapbox_full_address:
                        type: string
                        example: 'Paris, France'
                        description: 'The full address of the created journey.'
                        enum: []
                      mapbox_id:
                        type: string
                        example: 1234567890abcdefghij
                        description: 'The Mapbox ID of the created journey. For internal use only.'
                        enum: []
                      invite:
                        type: string
                        example: bf74e3c0-35d0-4854-9335-6ac6b9ce8956
                        description: 'The invite code of the created journey.'
                        enum: []
                      longitude:
                        type: number
                        example: 2.348392
                        description: 'The longitude of the created journey.'
                        enum: []
                      latitude:
                        type: number
                        example: 48.853495
                        description: 'The latitude of the created journey.'
                        enum: []
                      length:
                        type: integer
                        example: 8
                        description: 'The length of the journey in days.'
                        enum: []
                      id:
                        type: string
                        example: 9e807de1-4b8e-4cbd-9dc5-cf2addb6a89a
                        description: 'The ID of the created journey.'
                        enum: []
                      updated_at:
                        type: string
                        example: '2025-03-23T16:24:09.000000Z'
                        description: 'The last update date of the created journey.'
                        enum: []
                      created_at:
                        type: string
                        example: '2025-03-23T16:24:09.000000Z'
                        description: 'The creation date of the created journey.'
                        enum: []
                      creator:
                        type: string
                        example: null
                        description: 'The creator of the created journey if it is a template, currently not used here. Would be provided with username and display name or slug and name in the case of a business.'
                        enum: []
                    required:
                      - name
                      - destination
                      - from
                      - to
                      - mapbox_full_address
                      - mapbox_id
                      - invite
                      - longitude
                      - latitude
                      - length
                      - id
                      - updated_at
                      - created_at
                      - creator
                    description: 'The created journey.'
                required:
                  - message
                  - journey
        422:
          description: 'Validation error.'
          content:
            application/json:
              schema:
                type: object
                example:
                  message: 'The name field is required. (and 3 more errors)'
                  errors:
                    name:
                      - 'The name field is required.'
                    destination:
                      - 'The destination field is required.'
                    from:
                      - 'The from field is required.'
                    to:
                      - 'The to field is required.'
                properties:
                  message:
                    type: string
                    example: 'The name field is required. (and 3 more errors)'
                    description: 'The succes/error message.'
                    enum: []
                  errors:
                    type: object
                    properties:
                      name:
                        type: array
                        example:
                          - 'The name field is required.'
                        items:
                          type: string
                      destination:
                        type: array
                        example:
                          - 'The destination field is required.'
                        items:
                          type: string
                      from:
                        type: array
                        example:
                          - 'The from field is required.'
                        items:
                          type: string
                      to:
                        type: array
                        example:
                          - 'The to field is required.'
                        items:
                          type: string
                    description: 'The validation errors.'
                required:
                  - message
                  - errors
      tags:
        - Journey
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: 'Name of the journey.'
                  example: 'Seine-sational trip'
                  nullable: false
                destination:
                  type: string
                  description: 'Destination of the journey.'
                  example: Paris
                  nullable: false
                from:
                  type: string
                  description: 'Start date of the journey. Must be a valid date.'
                  example: '2021-01-01'
                  nullable: false
                to:
                  type: string
                  description: 'End date of the journey. Must be a valid date.'
                  example: '2021-01-10'
                  nullable: false
                invite:
                  type: string
                  description: 'Invite code of the journey. Must be a valid UUID.'
                  example: 123e4567-e89b-12d3-a456-426614174000
                  nullable: false
                mapbox_id:
                  type: string
                  description: 'Mapbox id of the journey.'
                  example: mapbox.123456
                  nullable: true
                mapbox_full_address:
                  type: string
                  description: 'Mapbox full address of the journey.'
                  example: 'Paris, France'
                  nullable: true
                template_id:
                  type: string
                  description: 'Template id of the journey. The <code>id</code> of an existing record in the journeys table.'
                  example: 123e4567-e89b-12d3-a456-426614174000
                  nullable: true
                calendar_activity_insert_mode:
                  type: string
                  description: 'Calendar activity insert mode of the journey if generating from a template. This field is required when <code>template_id</code> is present.'
                  example: direct
                  nullable: false
                  enum:
                    - direct
                    - smart
                    - pool
              required:
                - name
                - destination
                - from
                - to
...